change docs, move test function
