+++ torchmetrics/image/ergas.py
@@ -59,6 +59,7 @@ class ErrorRelativeGlobalDimensionlessSynthesis(Metric):
     preds: List[Tensor]
     target: List[Tensor]
     higher_is_better: bool = False
+    is_differentiable: bool = True
 
     def __init__(
         self,
